Simplify seeding setup
